Halfway fixed it, I honestly don't understand why it only loads 0s
I've checked and it isn't discarding DL either (although, I don't know 
if it's even *trying* to overwrite it - maybe not..?)

It's not setting the carry flag either, and it's setting ah to 00h, so 
it seems like it should have been successful, but it.. isn't

+++ Common/Kernel/Disk/Bios/Bios.c
@@ -35,7 +35,7 @@ const uint8 Int13Wrapper[] = {
 
 // (TODO - The function itself..?)
 
-typedef uint16 (*Fn_Int13Wrapper)(uint64 Lba, uint32 NumSectors, uint8 DriveNumber);
+typedef uint16 (*Fn_Int13Wrapper)(uint64 Lba, uint16 NumSectors, uint8 DriveNumber);
 static Fn_Int13Wrapper Call_Int13Wrapper = (Fn_Int13Wrapper)Int13Wrapper_Location;
